+++ public/preload.js
@@ -1,7 +1,7 @@
 const { contextBridge, ipcRenderer } = require("electron");
 
-// 일렉트론 API를 웹 콘텐츠에 노출
-contextBridge.exposeInMainWorld("electron", {
+// API 그룹 정의
+const captureAPI = {
   // 활성 창 목록 가져오기
   getActiveWindows: async () => {
     return await ipcRenderer.invoke("get-active-windows");
@@ -19,6 +19,24 @@ contextBridge.exposeInMainWorld("electron", {
     ipcRenderer.send("stop-capture");
   },
 
+  // 캡처 상태 이벤트 구독
+  onCaptureStatus: (callback) => {
+    console.log("Preload: onCaptureStatus 이벤트 구독 설정");
+    const captureStatusListener = (event, status) => {
+      callback(status);
+    };
+
+    ipcRenderer.on("capture-status", captureStatusListener);
+
+    // 컴포넌트 언마운트 시 이벤트 리스너 정리를 위한 함수 반환
+    return () => {
+      console.log("Preload: 캡처 상태 이벤트 구독 해제");
+      ipcRenderer.removeListener("capture-status", captureStatusListener);
+    };
+  },
+};
+
+const timelapseAPI = {
   // 타임랩스 생성
   generateTimelapse: (options) => {
     console.log("Preload: generateTimelapse 호출됨", options);
@@ -34,21 +52,9 @@ contextBridge.exposeInMainWorld("electron", {
       ipcRenderer.send("generate-timelapse", options);
     });
   },
+};
 
-  // 캡처 상태 이벤트 구독
-  onCaptureStatus: (callback) => {
-    const captureStatusListener = (event, status) => {
-      callback(status);
-    };
-
-    ipcRenderer.on("capture-status", captureStatusListener);
-
-    // 컴포넌트 언마운트 시 이벤트 리스너 정리를 위한 함수 반환
-    return () => {
-      ipcRenderer.removeListener("capture-status", captureStatusListener);
-    };
-  },
-
+const windowAPI = {
   // 창 제어 API
   minimize: () => {
     ipcRenderer.send("window:minimize");
@@ -70,4 +76,11 @@ contextBridge.exposeInMainWorld("electron", {
       ipcRenderer.send("window:is-maximized");
     });
   },
+};
+
+// 모든 API를 통합하여 노출
+contextBridge.exposeInMainWorld("electron", {
+  ...captureAPI,
+  ...timelapseAPI,
+  ...windowAPI,
 });
